Moved application initialization into EditorApplication class
